+++ packages/circuit-ui/styles/style-mixins.spec.tsx
@@ -273,6 +273,14 @@ describe('Style helpers', () => {
       );
     });
 
+    it('should match the snapshot when disabled', () => {
+      const { styles } = inputOutline({
+        theme: light,
+        disabled: true,
+      });
+      expect(styles).toMatchInlineSnapshot(`"box-shadow:0 0 0 1px #999;"`);
+    });
+
     it('should match the snapshot when invalid', () => {
       const { styles } = inputOutline({
         theme: light,