+++ packages/circuit-ui/styles/style-mixins.ts
@@ -324,8 +324,8 @@ export const hideScrollbar = (): SerializedStyles => css`
 `;
 
 /**
- * Visually communicates to the user that an input is hovered, focused,
- * or active.
+ * Visually communicates to the user that an element is hovered, focused, or
+ * active in the disabled, invalid, and warning states.
  */
 export const inputOutline = (
   args:
@@ -335,6 +335,9 @@ export const inputOutline = (
         disabled?: boolean;
         invalid?: boolean;
         hasWarning?: boolean;
+        /**
+         * @deprecated
+         */
         showValid?: boolean;
       },
 ): SerializedStyles => {